Changing name from astropile to mmu
