--- astropile/benchmark/__init__.py
@@ -1,2 +1,0 @@
-from astropile.benchmark.dataset import AstroPile, CrossMatchedAstroPile
-from astropile.benchmark.models.image import ConvolutionalModel
