--- astropile/benchmark/eval/__init__.py
@@ -1,1 +1,0 @@
-from .photo_z import PhotozEvalCallback
