--- astropile/benchmark/trainer.py
@@ -1,7 +1,0 @@
-from lightning.pytorch.cli import LightningCLI
-
-def trainer_cli():
-    return LightningCLI()
-
-if __name__ == "__main__":
-    trainer_cli()
