+++ benchmark_cfgs/image/regression_galaxy_photoz_resnet.yaml
@@ -1,5 +1,5 @@
 model:
-  class_path: astropile.benchmark.models.image.ConvolutionalModel
+  class_path: mmu.benchmark.models.image.ConvolutionalModel
   init_args:
     input_channels: 3
     output_size: 1
@@ -7,12 +7,12 @@ model:
     lr: 5e-4
     target: 'Z'
 data:
-  class_path: CrossMatchedAstroPile
+  class_path: CrossMatchedMMU
   init_args:
     left: 'desi'
     right: 'decals'
     right_config_name: 'stein_et_al_north'
-    local_astropile_root: '/mnt/ceph/users/polymathic/AstroPile_v1'
+    local_mmu_root: '/mnt/ceph/users/polymathic/MultimodalUniverse'
     cache_dir: '/home/flanusse/HF_cache2'
     batch_size: 256
     num_workers: 32
@@ -21,4 +21,4 @@ trainer:
   max_epochs: 10
   accelerator: gpu
   callbacks:
-  - class_path: astropile.benchmark.eval.PhotozEvalCallback
+  - class_path: mmu.benchmark.eval.PhotozEvalCallback
