+++ experimental_benchmark/galaxy10_decals/config.yaml
@@ -7,7 +7,7 @@ model:
 data:
   class_path: dataset.GZ10Dataset
   init_args:
-    dataset_path: '/mnt/ceph/users/polymathic/AstroPile_v1/gz10/gz10.py'
+    dataset_path: '/mnt/ceph/users/polymathic/MultimodalUniverse/gz10/gz10.py'
     batch_size: 256
     val_size: 0.2
 trainer:
