+++ mmu/benchmark/__init__.py
@@ -1,0 +1,2 @@
+from mmu.benchmark.dataset import MMU, CrossMatchedMMU
+from mmu.benchmark.models.image import ConvolutionalModel
