+++ mmu/benchmark/eval/__init__.py
@@ -1,0 +1,1 @@
+from .photo_z import PhotozEvalCallback
