+++ scripts/hsc/rusty.job
@@ -7,9 +7,9 @@
 
 module load modules/2.3-20240529 python
 
-source ~/venvs/astropile2/bin/activate
+source ~/venvs/mmu/bin/activate
 
 srun python build_parent_sample.py pdr3_dud_22.5.sql\
-        /mnt/ceph/users/polymathic/AstroPile_v1/hsc\
+        /mnt/ceph/users/polymathic/MultimodalUniverse/hsc\
         /mnt/ceph/users/polymathic/external_data/astro/hsc --nsplits 10 --num_processes 32
